==> notebooks/5a_generation_model_entmoot_training.ipynb <==
import os
import pandas as pd
os.chdir('..')
from src.models.entmoot_functions import *
from src.models.train_animation_predictor import *

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

### 1. Load and preprocess training data for optimizer fitting

#### Load data

X_train = pd.read_csv('data/path_selector/path_selector_train.csv')

X_train.head()

X_train = retrieve_m1_predictions(X_train)
X_train = retrieve_animation_midpoints(X_train, drop=True)

scaler = pickle.load(open(config.scaler_path, 'rb'))
X_train[config.sm_features] = scaler.transform(X_train[config.sm_features])

X_train.head()

#### Merge path animation vectors and labels

with open('data/surrogate_model/animation_label.pkl', 'rb') as f:
    animations = pickle.load(f)

animations.head()

# get correct filenames to match
animations['filename'] = animations['file'].apply(lambda row: row.split("_animation")[0])

# drop unnecessary columns
animations.drop(['file','order_id','path_prob','begin_value','animated_animation_ids','animated_order_ids','backend_mapping','logo_id','animation_number','alias','animation_file', 'time'], axis=1, inplace=True)

# map label names to label indexes and delete entries without rating
mapping = {'Very Bad': 0, 'Bad': 1, 'Okay': 2, 'Good': 3, 'Very Good': 4}
animations.replace({'label': mapping}, inplace=True)
animations = animations[animations['label'] != 'no_rating']
animations.reset_index(drop=True, inplace=True)

animations.head()

initial_data = animations.merge(X_train, how='left', on=['filename', 'animation_id'])

print(f"Before: Train: {initial_data.shape}.")
initial_data.dropna(inplace=True)
print(f"After: Train: {initial_data.shape}.")

# Drop unnecessary columns
initial_data.drop(['stroke_width','opacity','stroke_opacity','stroke_r','stroke_g','stroke_b','svg_stroke_r','diff_stroke_r','svg_stroke_g','diff_stroke_g','svg_stroke_b','diff_stroke_b'], axis=1, inplace=True)

initial_data.head()

#### Transform animation vector into multiple dataframe columns and change column ordering

initial_data[[f'an_vec_{i}' for i in range(12)]] = pd.DataFrame(initial_data['model_output'].tolist(), index=initial_data.index)

initial_data.drop(['model_output','animation_id','filename'], inplace=True, axis=1)

col_order = [f'an_vec_{i}' for i in range(12)] + [f'emb_{i}' for i in range(10)] + ['_'.join(['fill', ch]) for ch in ['r','g','b']] + ['_'.join(['svg_fill', ch]) for ch in ['r','g','b']] + ['_'.join(['diff_fill', ch]) for ch in ['r','g','b']] + ['rel_height','rel_width','rel_x_position','rel_y_position','rel_x_position_to_animations','rel_y_position_to_animations','nr_paths_svg','label']
initial_data = initial_data[col_order]

#### Encode labels into 4-binary labels

initial_data[[f'rating_{i}' for i in range(4)]] = pd.DataFrame(encode_classes(np.array(initial_data['label'])), index=initial_data.index)

initial_data.drop(['label'], axis=1, inplace=True)

initial_data.reset_index(drop=True, inplace=True)

initial_data.head()

initial_data.to_csv('data/entmoot_optimization/entmoot_initial_data_full.csv', index=False)

### 2. Train Bayesian optimizer

# Load subset of initial data (100 training examples)
initial_data = pd.read_csv('data/entmoot_optimization/entmoot_initial_data_reduced.csv')

initial_data.head()

X_train = initial_data.iloc[:, :-4]

# use 0-padding instead of padding with -1 
X_train.replace(to_replace=-1, value=0, inplace=True)

y_train = initial_data.iloc[:, -4:]

y_train = pd.Series(decode_classes(y_train.to_numpy()).flatten()) * -1

X_train = X_train.values.tolist()
y_train = y_train.values.tolist()

# Load surrogate model for function evaluations
func = SurrogateModelFNN()

# Fit optimizer to given initial data

optimizer = None
start_time = time.time()
optimizer = entmoot_fit(dimensions=func.get_bounds(), x0=X_train, y0=y_train, random_state=73, base_estimator="RF", std_estimator="L1DDP")
print("--- %s seconds ---" % (time.time() - start_time))

# Save fitted optimizer
with open(f'models/entmoot_optimizer_final.pkl', 'wb') as output:
    pickle.dump(optimizer, output, pickle.HIGHEST_PROTOCOL)

# Show example of animation prediction

# Load optimizer
with open("models/entmoot_optimizer_final.pkl", "rb") as f:
    optimizer = pickle.load(f)

# Predict animations for new path
path_vector = [9.0731e-02,  9.1726e-01, -2.2292e-01,
         1.0261e+00, -1.0310e+00,  3.8478e-01,  4.4583e-01,  3.7610e+00,
         9.0707e-02, -1.5281e+00, -1.1364e+00, -1.0705e+00, -8.6565e-01,
        -7.8171e-01, -9.8579e-01, -8.6640e-01, -6.1583e-01, -3.6518e-01,
        -2.7900e-01, -7.6306e-01,  2.0780e+00,  4.8552e-01, -1.5957e-01,
        -2.7360e-03, -1.3878e+00,  9.0921e-02]

start_time = time.time()
opt_x, opt_y,_ = entmoot_predict(optimizer, func, path_vector, n_calls=1)
print("--- %s seconds ---" % (time.time() - start_time))

print(f'Predicted animation vector: {opt_x}')
print('-'*60)
print(f'Predicted animation score: {-opt_y}')
