==> notebooks/6_logo_animation_pipeline.ipynb <==
# SVG Pipeline

import os
os.chdir("..")
from src.pipeline import * 

### Load Logo

logo = Logo(data_dir="data/uploads/logo_169.svg")  # previously: logo_181

### Print Logo Information

logo.print_logo_information()

### Create Logo Animations

%%time
logo.animate(model='backprop')

%%time
logo.animate(model='opt')
